% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_annotations.R
\name{get_colors_simplified}
\alias{get_colors_simplified}
\title{Get colors with simplified approach}
\usage{
get_colors_simplified(n_colors, palette_input)
}
\arguments{
\item{n_colors}{Number of colors needed}

\item{palette_input}{Either palette name (string) or color vector}
}
\value{
Vector of colors
}
\description{
Get colors with simplified approach
}
\keyword{internal}
